requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic_settings
python-dotenv
asyncpg